adjust colors
